rename spacebar peek to peek viewer
