+++ src/modules/peek/PeekViewer/FileTypeUtils.cpp
