+++ src/modules/peek/PeekViewer/PeekViewer.rc
@@ -52,7 +52,7 @@ END
 
 // Icon with lowest ID value placed first to ensure application icon
 // remains consistent on all systems.
-IDI_ICON1               ICON                    "spacebar.ico"
+IDI_ICON1               ICON                    "PeekViewer.ico"
 
 #endif    // English (United States) resources
 /////////////////////////////////////////////////////////////////////////////
